+++ lib/models/test_action.dart
@@ -64,70 +64,59 @@ class TestAction {
   });
 
   ///Performs the action
-  Future<void> performAction({Function? setAsDone, int? actionIndex}) async {
-    print(
-        'Performing ${actionType.toValue} action ${actionIndex != null ? actionName ?? '#$actionIndex' : ''}');
+  Future<void> performAction(
+      {Function(int)? setAsDone, int actionIndex = 0}) async {
+    debugPrint('Performing ${actionType.toValue} action $actionIndex');
     try {
       switch (actionType) {
-        case ActionType.Press:
+        case ActionType.press:
           await tester!.tap(finder!);
           break;
-        case ActionType.EnterText:
+        case ActionType.enterText:
           await tester!.enterText(finder!, enterText!);
           break;
-        case ActionType.PumpAndSettle:
+        case ActionType.pumpAndSettle:
           await tester!.pumpAndSettle();
           break;
-        case ActionType.FutureAwait:
-          print(
-              '** Future delayed by ${awaitDuration!.inSeconds}s -> ${awaitDuration!.inMilliseconds}ms');
-          await Future.delayed(awaitDuration!);
+        case ActionType.futureAwait:
+          await _executeFutureAwait();
           break;
-        case ActionType.CustomAction:
+        case ActionType.customAction:
           await customAction!();
           break;
-        case ActionType.Pump:
-          print('* Pumping for $pumpTime times');
-          bool pumpDone = false;
-          for (int i = 0; i < pumpTime; i++) {
-            await tester!.pump(Duration(milliseconds: 100));
-            if (i == pumpTime - 1) pumpDone = true;
-          }
-          do {
-            await Future.delayed(Duration(milliseconds: 100));
-          } while (!pumpDone);
-          if (setAsDone != null) setAsDone(actionIndex);
-          print('Action ${actionType.toValue} #$actionIndex done');
+        case ActionType.pump:
+          await _executePump(setAsDone, actionIndex);
           return;
-        case ActionType.Drag:
+        case ActionType.drag:
           await tester!.drag(finder!, dragOffset!);
           break;
       }
       if (executePumpAndSettle) {
-        print('Executing pumpAndSettle');
-        await tester!.pumpAndSettle(Duration(milliseconds: 500));
-      } else
-        print('* Avoiding the pumpAndSettle');
+        debugPrint('Executing pumpAndSettle');
+        await tester!.pumpAndSettle(const Duration(milliseconds: 500));
+      } else {
+        debugPrint('* Avoiding the pumpAndSettle');
+      }
     } catch (e, stacktrace) {
-      printError(e, actionIndex ?? 0, surplusMessage: '\n$stacktrace');
+      _printError(e, actionIndex, surplusMessage: '\n$stacktrace');
       return;
     }
     if (setAsDone != null) setAsDone(actionIndex);
-    print('Action ${actionType.toValue} #$actionIndex done');
+    debugPrint('Action ${actionType.toValue} #$actionIndex done');
     return;
   }
 
-  void printError(dynamic e, int actionIndex, {String? surplusMessage}) {
+  void _printError(dynamic e, int actionIndex, {String? surplusMessage}) {
     String firstMessage =
         '* Error occurred while performing ${actionType.toValue} action: ${actionName ?? '#' + actionIndex.toString()}';
     String secondMessage =
         '* ErrorType: ${e.runtimeType}\t Message: ${e.toString()}';
     String divider = ''.padLeft(firstMessage.length + 1, "=");
-    print(divider);
-    print(firstMessage);
-    print(secondMessage);
-    if (surplusMessage != null) print('* $surplusMessage');
-    print(divider);
+    debugPrint(divider);
+    debugPrint(firstMessage);
+    debugPrint(secondMessage);
+    if (surplusMessage != null) debugPrint('* $surplusMessage');
+    debugPrint(divider);
   }
 
   TestAction copyWith({
@@ -135,7 +124,7 @@ class TestAction {
     WidgetTester? tester,
     Finder? finder,
     String? enterText,
-    Duration? awaitFuture,
+    Duration? awaitDuration,
     Function? customAction,
     bool? executePumpAndSettle,
     Offset? dragOffset,
@@ -159,19 +148,42 @@ class TestAction {
   bool operator ==(Object other) {
     List<String> differences = [];
     if (runtimeType != other.runtimeType) differences.add('runTime Type');
-    if (!(other is TestAction)) return false;
+    if (other is! TestAction) return false;
     if (actionType != other.actionType) differences.add('ActionType');
     if (pumpTime != other.pumpTime) differences.add('pumpTime');
     if (enterText != other.enterText) differences.add('enterText');
     if (awaitDuration != other.awaitDuration) differences.add('awaitDuration');
-    if (executePumpAndSettle != other.executePumpAndSettle)
+    if (executePumpAndSettle != other.executePumpAndSettle) {
       differences.add('executePumpAndSettle');
+    }
     if (dragOffset != other.dragOffset) differences.add('dragOffset');
     if (actionName != other.actionName) differences.add('actionName');
-    if (differences.length > 0) print('Differences: $differences');
-    return differences.length == 0;
+    if (differences.isNotEmpty) debugPrint('Differences: $differences');
+    return differences.isEmpty;
   }
 
   @override
   int get hashCode => actionType.toValue.hashCode;
+
+  ///Executes the pump
+  Future<void> _executePump(Function(int)? setAsDone, int? actionIndex) async {
+    debugPrint('* Pumping for $pumpTime times');
+    bool pumpDone = false;
+    for (int i = 0; i < pumpTime; i++) {
+      await tester!.pump(const Duration(milliseconds: 100));
+      if (i == pumpTime - 1) pumpDone = true;
+    }
+    do {
+      await Future.delayed(const Duration(milliseconds: 100));
+    } while (!pumpDone);
+    if (setAsDone != null) setAsDone(actionIndex ?? 0);
+    debugPrint('Action ${actionType.toValue} #$actionIndex done');
+  }
+
+  ///Executes the await on the future
+  Future<void> _executeFutureAwait() async {
+    debugPrint(
+        '** Future delayed by ${awaitDuration!.inSeconds}s -> ${awaitDuration!.inMilliseconds}ms');
+    await Future.delayed(awaitDuration!);
+  }
 }