+++ lib/models/test_actions.dart
@@ -1,5 +1,6 @@
 import 'dart:async';
 
+import 'package:flutter/foundation.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:test_actions/models/test_action.dart';
 
@@ -16,12 +17,13 @@ class TestActions {
   dynamic get tester => _tester;
 
   TestActions({List<TestAction>? actions, WidgetTester? tester}) {
-    if (actions != null && actions.length > 0)
-      actions.forEach((element) {
+    if (actions != null && actions.isNotEmpty) {
+      for (var element in actions) {
         addAction(element);
-      });
-    else
+      }
+    } else {
       _actions = [];
+    }
     _doneActions = [];
     if (tester != null) _tester = tester;
   }
@@ -41,8 +43,8 @@ class TestActions {
     Timer.periodic(failsAt, (timer) {
       throw Exception('Test not completed after 5 minutes');
     });
-    if (_actions.length == 0) {
-      print('No actions to do');
+    if (_actions.isEmpty) {
+      debugPrint('No actions to do');
       return;
     }
     await _recursivePerformAction(_actions, 0);
@@ -52,21 +54,24 @@ class TestActions {
   //Performs the actions recursively
   Future<void> _recursivePerformAction(
       List<TestAction> actionsList, int currentActionIndex) async {
-    if (!_doneActions[currentActionIndex])
+    if (!_doneActions[currentActionIndex]) {
       await actionsList.first.performAction(
           actionIndex: currentActionIndex,
           setAsDone: (_) => _doneActions[currentActionIndex] = true);
+    }
     actionsList.removeAt(0);
     //Performs the action if there's any
-    if (actionsList.length > 0)
+    if (actionsList.isNotEmpty) {
       await _recursivePerformAction(actionsList, currentActionIndex + 1);
+    }
   }
 
   ///Performs a single action.
   Future<void> performActionAt(int index) async {
-    if (index > _actions.length)
+    if (index > _actions.length) {
       throw Exception(
           'Out of bound exception, the selected index [$index] is greater then the list [${_actions.length}]');
+    }
     if (!_doneActions[index]) {
       await _actions[index]
           .performAction(setAsDone: _setActionAsDone, actionIndex: index);
@@ -75,7 +80,7 @@ class TestActions {
 
   ///Adds an action at the end of the others.
   ///
-  /// If the [WidgetTester] has been set, will be used instead of the one in the [TestAction]
+  ///If the [WidgetTester] has been set, will be used instead of the one in the [TestAction]
   void addAction(TestAction action) {
     _actions.add(_tester != null ? action.copyWith(tester: _tester) : action);
     _doneActions.add(false);
@@ -83,24 +88,24 @@ class TestActions {
 
   ///Adds a [List] of [TestAction]
   void addActionsAll(List<TestAction> multipleActions) {
-    multipleActions.forEach((element) {
+    for (var element in multipleActions) {
       addAction(element);
-    });
-    print('* Added ${_actions.length} actions');
+    }
+    debugPrint('* Added ${_actions.length} actions');
   }
 
   ///Adds a [TestAction] at [int] index
   void addActionAt(int index, TestAction action) {
     _actions.insert(
         index, _tester != null ? action.copyWith(tester: _tester) : action);
-    print('* Added a new action at index $index');
+    debugPrint('* Added a new action at index $index');
   }
 
   void _setActionAsDone(int index) => _doneActions[index] = true;
 
   ///Resets the actions and the [WidgetTester]
   void resetActions() {
-    print('** Actions reset **');
+    debugPrint('** Actions reset **');
     _actions = [];
     _doneActions = [];
   }